Make the old minixST plat work, or at least build.

+++ mach/m68k2/as/mach1.c
@@ -55,9 +55,9 @@ extern int	curr_instr;
 #define	PUTL		0x80
 
 #ifndef extern
-extern short	eamode[];
+extern int	eamode[];
 #else
-short	eamode[] = {
+int	eamode[] = {
 /* 00A */	DTA        |ALT,
 /* 01A */	            ALT,
 /* 02A */	DTA|MEM|CTR|ALT,
